Adding NSE Dataplane clean
