+++ plugins/finalizer/finalizer_queue.go
@@ -18,6 +18,8 @@ import (
 	"fmt"
 	"reflect"
 
+	dataplaneutils "github.com/ligato/networkservicemesh/pkg/dataplane/utils"
+	"github.com/ligato/networkservicemesh/pkg/nsm/apis/simpledataplane"
 	"k8s.io/api/core/v1"
 	"k8s.io/client-go/tools/cache"
 	"k8s.io/client-go/util/workqueue"
@@ -96,6 +98,12 @@ func cleanUpNSE(plugin *Plugin, pod *v1.Pod) error {
 	plugin.Log.Infof("all channels advertised by NSE %s/%s were deleted", pod.ObjectMeta.Namespace, pod.ObjectMeta.Name)
 
 	// Step 4 Clean up dataplane
+	if err := dataplaneutils.CleanupPodDataplane(pod.ObjectMeta.Name, pod.ObjectMeta.Namespace, simpledataplane.NSMPodType_NSE); err != nil {
+		// NSE pod is about to be deleted as such there is no reason to fail cleanUpNSE, even if
+		// dataplane cleanup failed, simply print an error message is sufficient
+		plugin.Log.Errorf("failed to clean up pod %s/%s dataplane with error: %+v, please review dataplane controller log if further debugging is required",
+			pod.ObjectMeta.Namespace, pod.ObjectMeta.Name, err)
+	}
 
 	return nil
 }
